+++ R/retailer.R
@@ -36,6 +36,8 @@ retailer <- function(retailer_id = NA) {
 #' Products for a specific retailer
 #'
 #' @param retailer_id A retailer ID.
+#' @param name Filter by product name (treated as a regular expression).
+#' @param brand Filter by product brand (treated as a regular expression).
 #' @param ... Arguments passed through to \code{paginate()}.
 #'
 #' @return Details of all products for a specific retailer as a \code{data.frame}.
@@ -45,11 +47,21 @@ retailer <- function(retailer_id = NA) {
 #' # Get products for a specific retailer.
 #' \dontrun{
 #' retailer_products(1)
+#' retailer_products(9, name = "Nescafe")
 #' }
-retailer_products <- function(retailer_id, ...) {
-  products <- paste0(base_url(), "retailer/%d/product") %>%
-    sprintf(retailer_id) %>%
-    paginate(...)
+retailer_products <- function(retailer_id, name = NA, brand = NA,...) {
+  url <- paste0(base_url(), "retailer/%d/product") %>%
+    sprintf(retailer_id)
+
+  if (!is.na(name)) {
+    url <- param_set(url, key = "name", value = name)
+  }
+
+  if (!is.na(brand)) {
+    url <- param_set(url, key = "brand", value = brand)
+  }
+
+  products <- paginate(url, ...)
 
   if (nrow(products)) {
     products %>% rename(product_id = id)