+++ test.R
@@ -11,6 +11,8 @@ set_api_key(API_KEY)
 get_api_key()
 
 retailer_products(9, verbose = TRUE)
+retailer_products(9, name = "nescafe")
+retailer_products(10, brand = "nivea")
 product(546082) %>% select(-retailer_id, -url)
 product_prices(546082)
 product(531589) %>% select(-retailer_id, -brand, -url)