+++ NEWS.md
@@ -1,3 +1,8 @@
+# retail 0.0.7
+
+* Add filter by product name in retailer_products().
+* Add filter by product brand in retailer_products().
+
 # retail 0.0.6
 
 * Return 'model' field in retailer_products() and product().